Add CreateOrUpdate type and refactor repository functions
Added a new type `CreateOrUpdate` in `Definitions.Request` and
`Definitions.Chat` modules to handle both creation and updating
of `PassportsGroup` and `ChatSubscription` respectively. Updated
the `Request.Operation` and `Chat.Operation` types to include the
new `CreateOrUpdate` case.

Modified the `create` functions in `FileSystemRepository.fs` and
`InMemoryRepository.fs` to handle the new `CreateOrUpdate` type.
Added new `createOrUpdate` functions in these files to handle the
logic for creating or updating `PassportsGroup` and `ChatSubscription`.

Updated the `execute` functions in `FileSystemRepository.fs` and
`InMemoryRepository.fs` to call the new `createOrUpdate` functions.
Updated `Message.fs` and `Repository.fs` to use the new
`createOrUpdate` functions.

Refactored the `create`, `update`, and `delete` functions in
`FileSystemRepository.fs` and `InMemoryRepository.fs` to use the
term `definition` instead of `create`, `update`, or `delete` for
better consistency and clarity.